Removed the topic models from the PM-classifier.
They improve the CrossVal result only by .3 ACC. This was tested
on Erik's servers with the Topic Model that is referenced hard-coded
in the CrossVal class. The model is over 300MB big which is why it is
NOT CONTAINED in the Git repository. But it also doesn't seem to be
required. The crossval ACC is slightly over 75%.